Add Warning component to Shared component output
