+++ components/Shared/index.js
@@ -22,6 +22,7 @@ export { default as LoadingScreen } from './LoadingScreen'
 export { default as NetworkSwitcherGlyph } from './NetworkSwitcherGlyph'
 export { default as StepHeader } from './StepHeader'
 export { default as ErrorView } from './Error'
+export { default as Warning } from './Warning'
 export * from './Link'
 export * from './Menu'
 export * from './Icons'
